requirements.txt
torch
pyzmq
requests
numpy
pyre-extensions
cloudpickle
torchx-nightly
